Remove AssemblyInfo
We don't use it .net core. In all AssemblyInfo.cs are set 3 properties. AssemblyTitle and AssemblyProduct is the same as project name, so we don't need to those two. AssemblyCopyright is the same in all AsemblyInfo.cs, so we will set it up in one place in Directory.Build.Props.